requirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
Django==3.2
djangorestframework==3.12.4
idna==2.10
oauthlib==3.1.0
psycopg2-binary==2.8.6
PySocks==1.7.1
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
tweepy==3.10.0
urllib3==1.26.4
